feat(view_activities): use GtkListView for endless scrolling
The old layout with the ToggleButton group took up too much
horizontal space to properly fit on a phonescreen. Considering
that and that users who want to see older entries have to
scroll down anyway, we might as well remove the ToggleButtons
and replace them with a ListView which makes it somewhat cheap
to see all entries.

fixes #121