widgets: use adw::StyleManager and Widget::color over style_context.
gtk::StyleContext is marked as deprecated and will be removed for
gtk5. The Accent color can be obtained through adw::StyleManager. The
foreground color can be obtained through Widget::color().

This uses those colors and colors with reduced alpha to replace the hardcoded
lightblue and and insensitive_fg_color. Style context is still used
for now in one instance to get the success color, with the deprecation
warning maked as allowed.

Closes #175