gtk: port plugins and data_add_dialog to adw::NavigationPage.
The gtk::Dialog was deprecated and the details pages are better
handled with adw::NavigationPage than the custom code.

This ports the custom details pages and add_data_dialog to
adw::NavigationPage.
The joined details dialog is split into two pages. The joined add
button is removed and replaced by two circular buttons, as planned for
the new design (see issue #178).

This also temporarily removes the Disable Plugin and add buttons for
details pages, they will be re-added with new code in the next commit.